Switch to pessimistic versioning to avoid future breakage.
I have historically avoided pessimistic versioning because I hated for a
piece of software to be running old, possible insecure code just because it
*might* be broken in a new version. I figured the new breakage would reveal
itself quickly.

Decideded to give pessimistic versioning a try for this project. This should
avoid any breakage but it means I need to periodically explicitly update to
new versions to ensure I am not running on old crufty code.