Allow more flexible ActiveSupport versioning.
In general I still support the strategy of 1917f2bcd. I do not think
you should add this tool to your Gemfile and as it increases the chance
of version conflicts unnecessarily.

But since we are compatible with ActiveSupport 3.2 and ActiveSupport is
so common I have decided to loosen up the restriction in that one
place.